requirements.txt
accelerate==0.30.1
compel==2.0.2
diffusers==0.27.2
lpips==0.1.4
matplotlib==3.9.0
numpy==1.26.3
pandas==2.2.2
pillow==10.2.0
piq==0.8.0
rembg==2.0.56
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
statsmodels==0.14.2
tokenizers==0.19.1
torch==2.3.0+cu118
torchaudio==2.3.0+cu118
torchvision==0.18.0+cu118
tqdm==4.66.4
transformers==4.40.2
webcolors==1.13
